LESmain.luaを分割しシェル依存排除・グローバル変数汚染を修正
## 1. LESmain.lua (1908行) を7つのモジュールに分割
- menus/plugin.lua: メニュー構築 (buildPluginMenu, clearcategories等)
- lifecycle/reload.lua: リロード・チート (reloadLES, cheatmenu等)
- shortcuts/macros.lua: キーボードマクロ (quickmacro eventtap等)
- shortcuts/rightclick.lua: 右クリック処理 (firstRightClick等)
- shortcuts/piano.lua: ピアノロールマクロ (modifierHandler等)
- vst/shortcuts.lua: VST固有ショートカット (Kick 2 undo/redo)
- tracking/timer.lua: 時間計測 (coolfunc, timerfunc等)
- lifecycle/appwatch.lua: アプリ監視 (enablemacros, disablemacros等)
- LESmain.luaはスリムなオーケストレーターに

## 2. helpers.luaのシェル依存をLua/Hammerspoon APIに置換
- ShellCopy → io.open/read/write
- ShellCreateDirectory → hs.fs.mkdir (再帰対応)
- ShellOverwriteFile → io.open("w")
- ShellConcatenateFile → io.open("a")
- ShellCreateEmptyFile → io.open("w") + close
- ShellDeleteFile → os.remove / hs.fs.rmdir
- module.luaのls→ioIsFilePresentに置換
- ShellExec/ShellRecursiveCopy/ShellNSOpenは代替なしのため維持

## 3. グローバル変数汚染の修正
- string標準ライブラリを上書きするバグを修正 (menus/plugin.lua)
- filepaths.luaとutil/string.luaのstring引数名をpath/strに変更
- buildPluginMenu内の暗黙グローバルをlocal化
- module.lua内のネスト関数をlocal functionに変更
- proccom.luaのfetchInnerTitleをlocal化
- settings.luaのvalidateValueをlocal化
- buplicate shortcutのループ化 (7/8回の重複削除)
- clear track macroのループ化 (12回のキーストローク重複削除)
- requesttime()の時間計算を簡潔化
